scripts/requirements.txt
pandas>=1.5.0
numpy>=1.21.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
openpyxl>=3.0.0
